==> FINAL/Report/scripts/00_Report.qmd <==
---
title: "Chemical Sample Classification Report"
subtitle: ""
author: "Petter Lövehagen"
date: "09 May 2023"

RVersion: 3.6.0 
RStudio: 2023.03.0+386 Cherry Blossom Release
Platform: "x86_64-w64-mingw32/x64 (64-bit)"
OS: "Windows 11 SE"

format: 
  docx: default
  html:
    toc: true
    toc-depth: 2
    toc-title: "Contents"
    toc-location: left
    code-fold: true
    code-line-numbers: true
    theme: spacelab
    html-math-method: mathjax
    highlight: "tango"
    #css: "../data/custom.css"
#prefer-docx: true
prefer-html: true
lang: "en-GB"
bibliography: "../data/AS_CW_References.bib"
csl: "../data/harvard-university-of-the-west-of-england.csl"

knitr:
  opts_chunk:
    include: false
    #include: true
    tidy: true
    echo: false
    #echo: true
    warning: false
    message: false
    #comment: true
    error: true
    fig.align: left
    out.width: '90%'
    fig.width: 6
    fig.height: 4
---

#### <!--# Set up R script  -->

<!--# The script installs/loads packages needed for the Report. If a section is rendered separately, it only required libraries.  See R file for details. -->

<!--# The script sets a seed so that analysis and results can be reproduced. -->

<!--# The script sets table rendering defaults. -->

<!--# The complete report takes upto 15 minutes to compile -->

```{r packages}
file = "project"
# load/install R packages/libraries
source("R/chem_report_load.R")

# what is currently loaded/attached:
#sessionInfo()


```

```{r intro, child="01_Introduction.qmd"}
#add introductory section

```

\newpage

```{r data, child="02_Data.qmd"}
#add data section

```

\newpage

```{r eda, child="03_EDA.qmd"}
#add EDA section

```

\newpage

```{r dimred, child="04_Model_DimRed.qmd"}
#add Dimension Reduction section

```

\newpage

```{r cluster, child="05_Model_Cluster.qmd"}
#add Cluster section

```

\newpage

```{r child="06_Model_Classifier.qmd"}

#add Classifier section

```

\newpage

```{r child="07_Findings.qmd"}

#add Classifier section

```
